feat: modularize electron
