+++ electron/src/shared.ts
@@ -1,0 +1,20 @@
+export const shared: {
+    USER: {
+        online: boolean,
+        accounts: Array<{
+            pubkey: any;
+            caip: string;
+        }>,
+        balances: any[]
+    },
+    SIGNED_TX: any,
+    eventIPC: {}
+} = {
+    USER: {
+        online: false,
+        accounts: [],
+        balances: []
+    },
+    SIGNED_TX: null,
+    eventIPC: {}
+}
